+++ src/mako/chrono/TimeInterface.php
@@ -166,6 +166,8 @@ interface TimeInterface extends DateTimeInterface
 	/**
 	 * Returns a formatted date string according to current locale settings.
 	 *
+	 * @deprecated
+	 *
 	 * @param  string $format Date format
 	 * @return string
 	 */