+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 * Added the following methods to the `mako\utility\Arr` class:
 	- `Arr::toObject()`
 
+#### Deprecations
+
+* Deprecated the `TimeInterface::formatLocalized()` method.
+
 --------------------------------------------------------
 
 ### 8.1.5, 9.0.3 <small>(2023-01-21)</small>